+++ ui-ngx/src/app/modules/home/components/rule-node/external/azure-iot-hub-config.component.html
@@ -38,6 +38,7 @@
       {{ 'rule-node-config.device-id-required' | translate }}
     </mat-error>
   </mat-form-field>
+  <tb-mqtt-version-select formControlName="protocolVersion" subscriptSizing="fixed"></tb-mqtt-version-select>
   <mat-accordion>
     <mat-expansion-panel class="tb-mqtt-credentials-panel-group">
       <mat-expansion-panel-header>
